python/flask-websocket-project/requirements.txt
Flask
Flask-SocketIO
eventlet
redis
python-dotenv
flasgger